# Trying to make IPFix parsing faster 24% speedup