--- a/parser_fuzz.go
+++ b/parser_fuzz.go
@@ -1,3 +1,4 @@
+//go:build gofuzz
 // +build gofuzz
 
 package ipfix