What I did:

- Change the uint16Bitmask from a byte array inside a struct to just a byte array (**6% speedup**)
- Remove the slice object stuff from the hot path, just operate and check the slice directly (**18% speedup**)
- Add actions tests


This won't make any HUGE differences, but 24% is 24%...


## Before
```
GOMAXPROCS=1 go test -bench=BenchmarkFilterWalk -count=10
goos: linux
goarch: amd64
pkg: github.com/gravwell/ipfix
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkFilterWalk       460162              2719 ns/op        236978538.83 MB/s
BenchmarkFilterWalk       462705              2688 ns/op        241003555.30 MB/s
BenchmarkFilterWalk       461238              2726 ns/op        236840863.71 MB/s
BenchmarkFilterWalk       458752              2728 ns/op        235405923.47 MB/s
BenchmarkFilterWalk       460761              2710 ns/op        238067220.26 MB/s
BenchmarkFilterWalk       448224              2684 ns/op        233786313.57 MB/s
BenchmarkFilterWalk       459219              2730 ns/op        235511427.96 MB/s
BenchmarkFilterWalk       455536              2692 ns/op        236927802.73 MB/s
BenchmarkFilterWalk       463518              2680 ns/op        242156420.87 MB/s
BenchmarkFilterWalk       465410              2683 ns/op        242839800.41 MB/s
PASS
ok      github.com/gravwell/ipfix       15.616s
```

## After
```
GOMAXPROCS=1 go test -bench=BenchmarkFilterWalk -count=10
goos: linux
goarch: amd64
pkg: github.com/gravwell/ipfix
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkFilterWalk       554418              2102 ns/op        369241974.62 MB/s
BenchmarkFilterWalk       584043              2128 ns/op        384192021.31 MB/s
BenchmarkFilterWalk       568914              2190 ns/op        363707501.88 MB/s
BenchmarkFilterWalk       563606              2154 ns/op        366384547.92 MB/s
BenchmarkFilterWalk       583837              2088 ns/op        391551878.59 MB/s
BenchmarkFilterWalk       507873              2166 ns/op        328224198.60 MB/s
BenchmarkFilterWalk       582630              2082 ns/op        391840488.59 MB/s
BenchmarkFilterWalk       570936              2089 ns/op        382681439.99 MB/s
BenchmarkFilterWalk       578288              2084 ns/op        388495258.11 MB/s
BenchmarkFilterWalk       577342              2096 ns/op        385571534.90 MB/s
PASS
ok      github.com/gravwell/ipfix       15.257s

```